pyproject.toml
[project]
name = "fundica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.67",
    "datasets>=4.4.1",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pypandoc-binary>=1.16",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
